Hook in the new diary entry notification
When it's time for a user to enter a new notification dairy entry (which is
roughly once every 10 notifications), the applet will take a screenshot and
prompt the user to enter a new diary entry.

I still need to log the actual notification information to the disk, but this
is the initial steps of that work.