Don't process notifications twice
When a notification is created, it also seems to be updated directly after
being created. Because of this, we were incrementing the notification count by
two instead of one.

We don't handle notification creation anymore, in order to fix that. There was
nothing different about the two code paths anyways.